Use multiprocessing.Connection instead of queue.SimpleQueue
This is so that a future commit can use select.select() in the event consumer
watch both Connections at once.

To make this work, I've updated the hypothesis test to listen to the child
connection for expected events.  It's only expecting to see PredictionInput and
Cancel events (outside of worker shutdown) so it's pretty manageable to listen
on the connection for these events.

By listening for the PredictionInput event, we can ensure that the predict()
rule only returns once the prediction has actually been sent to the child by the
event consumer thread.  This avoids a race we were seeing where we sent a Done()
event before the worker had sent the PredictionInput event, wiping out the
prediction_in_flight state and resulting the event consumer seeing None when
looking up the state.

This also updates the cancel() functionality to go via the event consumer,
rather than have two different threads that can send messages to the child
worker.